# shared/openapi.json
{
  "components": {
    "schemas": {
      "AgentFinding": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "severity": {
            "enum": [
              "info",
              "warn",
              "risk"
            ],
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "severity",
          "message"
        ],
        "title": "AgentFinding",
        "type": "object"
      },
      "AgentResult": {
        "properties": {
          "findings": {
            "items": {
              "$ref": "#/components/schemas/AgentFinding"
            },
            "title": "Findings",
            "type": "array"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "score": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score",
            "type": "integer"
          }
        },
        "required": [
          "score",
          "label"
        ],
        "title": "AgentResult",
        "type": "object"
      },
      "AgentResultMap": {
        "properties": {
          "budgetAgent": {
            "$ref": "#/components/schemas/AgentResult"
          },
          "decisionAgent": {
            "$ref": "#/components/schemas/AgentResult"
          },
          "impulseAgent": {
            "$ref": "#/components/schemas/AgentResult"
          },
          "priceAgent": {
            "$ref": "#/components/schemas/AgentResult"
          },
          "reviewAgent": {
            "$ref": "#/components/schemas/AgentResult"
          }
        },
        "required": [
          "reviewAgent",
          "priceAgent",
          "budgetAgent",
          "impulseAgent",
          "decisionAgent"
        ],
        "title": "AgentResultMap",
        "type": "object"
      },
      "AnalyzeRequest": {
        "properties": {
          "platform": {
            "title": "Platform",
            "type": "string"
          },
          "priceHistory": {
            "items": {
              "$ref": "#/components/schemas/PriceHistoryPoint"
            },
            "title": "Pricehistory",
            "type": "array"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/Review"
            },
            "title": "Reviews",
            "type": "array"
          },
          "session": {
            "$ref": "#/components/schemas/SessionContext"
          },
          "userBudget": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserBudget"
              },
              {
                "type": "null"
              }
            ]
          },
          "userId": {
            "title": "Userid",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "platform",
          "product",
          "session"
        ],
        "title": "AnalyzeRequest",
        "type": "object"
      },
      "AnalyzeResponse": {
        "properties": {
          "agents": {
            "$ref": "#/components/schemas/AgentResultMap"
          },
          "decision": {
            "enum": [
              "green",
              "yellow",
              "red"
            ],
            "title": "Decision",
            "type": "string"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "recommendedAction": {
            "title": "Recommendedaction",
            "type": "string"
          },
          "riskScore": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Riskscore",
            "type": "integer"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "decision",
          "riskScore",
          "summary",
          "reasons",
          "agents",
          "recommendedAction"
        ],
        "title": "AnalyzeResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "PriceHistoryPoint": {
        "properties": {
          "date": {
            "title": "Date",
            "type": "string"
          },
          "price": {
            "title": "Price",
            "type": "number"
          }
        },
        "required": [
          "date",
          "price"
        ],
        "title": "PriceHistoryPoint",
        "type": "object"
      },
      "PriceObservationIn": {
        "description": "Payload from the extension on every product-page load.",
        "properties": {
          "currency": {
            "default": "TRY",
            "enum": [
              "TRY",
              "USD",
              "EUR"
            ],
            "title": "Currency",
            "type": "string"
          },
          "price": {
            "exclusiveMaximum": 10000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Price",
            "type": "number"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "url": {
            "maxLength": 1024,
            "minLength": 8,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "price"
        ],
        "title": "PriceObservationIn",
        "type": "object"
      },
      "PriceObservationOut": {
        "description": "Acknowledgement returned to the extension.",
        "properties": {
          "canonicalUrl": {
            "title": "Canonicalurl",
            "type": "string"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "platform": {
            "title": "Platform",
            "type": "string"
          },
          "storedAt": {
            "title": "Storedat",
            "type": "string"
          }
        },
        "required": [
          "canonicalUrl",
          "platform",
          "storedAt"
        ],
        "title": "PriceObservationOut",
        "type": "object"
      },
      "Product": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "currency": {
            "default": "TRY",
            "enum": [
              "TRY",
              "USD",
              "EUR"
            ],
            "title": "Currency",
            "type": "string"
          },
          "imageUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imageurl"
          },
          "originalPrice": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Originalprice"
          },
          "price": {
            "title": "Price",
            "type": "number"
          },
          "rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "reviewCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewcount"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "title",
          "price",
          "category",
          "url"
        ],
        "title": "Product",
        "type": "object"
      },
      "Review": {
        "properties": {
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "rating": {
            "title": "Rating",
            "type": "number"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "rating",
          "text",
          "date"
        ],
        "title": "Review",
        "type": "object"
      },
      "SessionContext": {
        "properties": {
          "clickSpeedMs": {
            "title": "Clickspeedms",
            "type": "number"
          },
          "currentHour": {
            "maximum": 23.0,
            "minimum": 0.0,
            "title": "Currenthour",
            "type": "integer"
          },
          "purchasesToday": {
            "minimum": 0.0,
            "title": "Purchasestoday",
            "type": "integer"
          },
          "searchedBefore": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Searchedbefore"
          },
          "timeOnPageSeconds": {
            "title": "Timeonpageseconds",
            "type": "number"
          }
        },
        "required": [
          "timeOnPageSeconds",
          "clickSpeedMs",
          "currentHour",
          "purchasesToday"
        ],
        "title": "SessionContext",
        "type": "object"
      },
      "UserBudget": {
        "properties": {
          "categoryLimit": {
            "title": "Categorylimit",
            "type": "number"
          },
          "categorySpent": {
            "title": "Categoryspent",
            "type": "number"
          },
          "currency": {
            "default": "TRY",
            "enum": [
              "TRY",
              "USD",
              "EUR"
            ],
            "title": "Currency",
            "type": "string"
          },
          "monthlyLimit": {
            "title": "Monthlylimit",
            "type": "number"
          },
          "monthlySpent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monthlyspent"
          }
        },
        "required": [
          "monthlyLimit",
          "categoryLimit",
          "categorySpent"
        ],
        "title": "UserBudget",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Türk e-ticaret için 5 ajanlı satın alma analiz servisi. Fiyat geçmişi crowdsource veritabanından beslenir.",
    "title": "Tartı — Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/analyze-purchase": {
      "post": {
        "description": "Ürün, yorum, fiyat geçmişi, bütçe ve oturum bilgisini alır; yeşil/sarı/kırmızı bir karar ve 3 gerekçe döner. `priceHistory` boş ise sunucu, ürün URL'sine ait crowdsource veritabanından son 90 günü çeker.",
        "operationId": "analyze_purchase_api_analyze_purchase_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ürün satın alımını 5 ajanla analiz et"
      }
    },
    "/api/analyze-purchase/stream": {
      "post": {
        "description": "Her satır bağımsız bir JSON nesnesidir (`application/x-ndjson`). Sıra: 4 paralel sinyal ajanı için `node_finished`, ardından decision için `node_finished`, en sonda `verdict` (tam AnalyzeResponse). Hata olursa son satır `error` olur. Karar mantığı normal endpoint ile aynıdır; tek fark stream şeklidir.",
        "operationId": "analyze_purchase_stream_api_analyze_purchase_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Aynı analiz, ajan tamamlanma olaylarını NDJSON ile streamler"
      }
    },
    "/api/health": {
      "get": {
        "description": "Process is up. Cheap; always 200 if the FastAPI worker can handle a request at all. Use for load-balancer health checks.",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Health Api Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Liveness probe"
      }
    },
    "/api/price-observation": {
      "post": {
        "description": "Eklenti her ürün sayfası açılışında o anki fiyatı bu uca POST eder. Sunucu URL'yi normalize edip append-only saklar; aynı ürünün ileride yapılan analizleri bu geçmişten beslenir. Rate-limit: IP başına 60/dakika.",
        "operationId": "post_price_observation_api_price_observation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PriceObservationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceObservationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Crowdsource fiyat gözlemi kaydet"
      }
    },
    "/api/ready": {
      "get": {
        "description": "Process is up AND the database is reachable. Returns 503 if the DB connection fails, so an orchestrator can hold traffic off until migrations finish on first boot.",
        "operationId": "ready_api_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Ready Api Ready Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Readiness probe"
      }
    },
    "/api/user-budget": {
      "get": {
        "description": "(userId, category) çifti için saklanan bütçeyi döner. Eşleşen kayıt yoksa permissive bir varsayılan döner; budget_agent böylelikle her zaman skorlayacak bir veri bulur.",
        "operationId": "get_user_budget_api_user_budget_get",
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "required": true,
            "schema": {
              "title": "Userid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": true,
            "schema": {
              "title": "Category",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserBudget"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Kullanıcı bütçesini (kategori bazlı) getir"
      },
      "put": {
        "description": "(userId, category) için upsert. Aynı çiftte birden fazla satır olmaz.",
        "operationId": "put_user_budget_api_user_budget_put",
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "required": true,
            "schema": {
              "title": "Userid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": true,
            "schema": {
              "title": "Category",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserBudget"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserBudget"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Kullanıcı bütçesini (kategori bazlı) kaydet/güncelle"
      }
    }
  }
}